auto-spend-guard/auto-spend-guard/requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
pandas>=2.1.4
openai>=1.10.0
python-dotenv>=1.0.0
graphviz>=0.20.1
tiktoken>=0.5.2
streamlit>=1.28.0
plotly>=5.15.0
